requirements.txt
# --------- pytorch --------- #
torch>=1.11.0
torchvision>=0.12.0
pytorch-lightning>=1.6.0
torchmetrics>=0.8.0
einops

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0
hydra-submitit-launcher>=1.1.6
wandb

# --------- models --------- #
librosa
# coral-pytorch>=1.2.0
# need to run install.sh from `openspeech` github repo
# openspeech-core>=0.3.0
transformers>=4.17.0
# need to install from
# must install those dependencies
# https://github.com/facebookresearch/TimeSformer
# timesformer



# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
scikit-learn    # used in some callbacks
seaborn         # used in some callbacks
jupyterlab      # better jupyter notebooks
pudb            # debugger